Moved internal cone cast pointer to a separate file, cleaned up other unneeded dependencies
