+++ Assets/MRTK/Services/InputSystem/GazeConePointer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 330d8ae500807044880a751b83ccd919
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
