+++ Assets/MRTK/Services/InputSystem/GazeProvider.cs
@@ -115,7 +115,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 {
                     gazeInputSource = new BaseGenericInputSource("Gaze", sourceType: InputSourceType.Head);
                     //internalPointer.SetGazeInputSourceParent(gazeInputSource);
-                    internalConeCastPointer.SetGazeInputSourceParent(gazeInputSource);
+                    coneCastPointer.SetGazeInputSourceParent(gazeInputSource);
                 }
 
                 return gazeInputSource;
@@ -143,16 +143,16 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 //}
                 //else
                 //{
-                    if (internalConeCastPointer == null)
-                    {
-                        InitializeGazePointer();
-                    }
-                    return internalConeCastPointer;
+                if (coneCastPointer == null)
+                {
+                    InitializeGazePointer();
+                }
+                return coneCastPointer;
                 //}
             }
         }
         //private InternalGazePointer internalPointer = null;
-        private InternalConeCastPointer internalConeCastPointer = null;
+        private GazeConePointer coneCastPointer = null;
 
         /// <inheritdoc />
         public GameObject GazeCursorPrefab { private get; set; }
@@ -193,8 +193,8 @@ namespace Microsoft.MixedReality.Toolkit.Input
         private bool delayInitialization = true;
         private Vector3 lastHeadPosition = Vector3.zero;
 
-        private Vector3? overrideHeadPosition = null;
-        private Vector3? overrideHeadForward = null;
+        public Vector3? overrideHeadPosition { get; private set; }
+        public Vector3? overrideHeadForward { get; private set; }
 
         #region InternalGazePointer Class
 
@@ -357,212 +357,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
             }
         }
 
-        #endregion InternalGazePointer Class
-
-        private class InternalConeCastPointer : GenericPointer,
-        IMixedRealityQueryablePointer
-        {
-            private readonly Transform gazeTransform;
-            private readonly BaseRayStabilizer stabilizer;
-            private readonly GazeProvider gazeProvider;
-
-            public override IMixedRealityController Controller { get; set; }
-
-            private float pointerExtent;
-
-            /// <inheritdoc />
-            public override float PointerExtent
-            {
-                get => pointerExtent;
-                set => pointerExtent = value;
-            }
-
-            // Is the pointer currently down
-            private bool isDown = false;
-
-            // Input source that raised pointer down
-            private IMixedRealityInputSource currentInputSource;
-
-            // Handedness of the input source that raised pointer down
-            private Handedness currentHandedness = Handedness.None;
-
-            /// <summary>
-            /// Only for use when initializing Gaze Pointer on startup.
-            /// </summary>
-            internal void SetGazeInputSourceParent(IMixedRealityInputSource gazeInputSource)
-            {
-                InputSourceParent = gazeInputSource;
-            }
-
-            public float coneCastSphereRadius = 0.05f;
-            public float coneCastRange = 10.0f;
-            public float coneCastAngle = 1.5f;
-            public float coneCastDistanceWeight = 0.25f;
-            public float coneCastAngleWeight = 1.0f;
-            public float coneCastDistanceToCenterWeight = 0.5f;
-            public float coneCastAngleToCenterWeight = 0.0f;
-            public LayerMask castLayerMask;
-
-            // Values ultimately returned by SceneQuery
-            private RaycastHit coneCastHit;
-
-            public InternalConeCastPointer(GazeProvider gazeProvider, string pointerName, IMixedRealityInputSource inputSourceParent, LayerMask[] raycastLayerMasks, float pointerExtent, Transform gazeTransform, BaseRayStabilizer stabilizer)
-                        : base(pointerName, inputSourceParent)
-            {
-                this.gazeProvider = gazeProvider;
-                PrioritizedLayerMasksOverride = raycastLayerMasks;
-                this.pointerExtent = pointerExtent;
-                this.gazeTransform = gazeTransform;
-                this.stabilizer = stabilizer;
-                IsInteractionEnabled = true;
-            }
-
-            private static readonly ProfilerMarker OnPreSceneQueryPerfMarker = new ProfilerMarker("[MRTK] InternalConeCastPointer.OnPreSceneQuery");
-
-            /// <inheritdoc />
-            public override void OnPreSceneQuery()
-            {
-                using (OnPreSceneQueryPerfMarker.Auto())
-                {
-                    Vector3 newGazeOrigin;
-                    Vector3 newGazeNormal;
-
-                    if (gazeProvider.IsEyeTrackingEnabledAndValid)
-                    {
-                        gazeProvider.gazeInputSource.SourceType = InputSourceType.Eyes;
-                        newGazeOrigin = gazeProvider.LatestEyeGaze.origin;
-                        newGazeNormal = gazeProvider.LatestEyeGaze.direction;
-                    }
-                    else
-                    {
-                        gazeProvider.gazeInputSource.SourceType = InputSourceType.Head;
-
-                        if (gazeProvider.UseHeadGazeOverride && gazeProvider.overrideHeadPosition.HasValue && gazeProvider.overrideHeadForward.HasValue)
-                        {
-                            newGazeOrigin = gazeProvider.overrideHeadPosition.Value;
-                            newGazeNormal = gazeProvider.overrideHeadForward.Value;
-                            // Reset values in case the override source is removed
-                            gazeProvider.overrideHeadPosition = null;
-                            gazeProvider.overrideHeadForward = null;
-                        }
-                        else
-                        {
-                            newGazeOrigin = gazeTransform.position;
-                            newGazeNormal = gazeTransform.forward;
-                        }
-
-                        // Update gaze info from stabilizer
-                        if (stabilizer != null)
-                        {
-                            stabilizer.UpdateStability(MixedRealityPlayspace.InverseTransformPoint(newGazeOrigin), MixedRealityPlayspace.InverseTransformDirection(newGazeNormal));
-                            newGazeOrigin = MixedRealityPlayspace.TransformPoint(stabilizer.StablePosition);
-                            newGazeNormal = MixedRealityPlayspace.TransformDirection(stabilizer.StableRay.direction);
-                        }
-                    }
-
-                    Vector3 endPoint = newGazeOrigin + (newGazeNormal * pointerExtent);
-                    Rays[0].UpdateRayStep(ref newGazeOrigin, ref endPoint);
-
-                    coneCastHit = ConeCastExtension.ConeCastBest(newGazeOrigin, Vector3.Normalize(newGazeNormal), coneCastSphereRadius, coneCastRange, coneCastAngle, PrioritizedLayerMasksOverride[0], coneCastDistanceWeight, coneCastAngleWeight, coneCastDistanceToCenterWeight, coneCastAngleToCenterWeight);
-                    Debug.Log(coneCastHit.collider?.name);
-                }
-            }
-
-            // Returns the hit values cached by the queryBuffer during the prescene query step
-            public bool OnSceneQuery(LayerMask[] prioritizedLayerMasks, bool focusIndividualCompoundCollider, out MixedRealityRaycastHit hitInfo)
-            {
-                PrioritizedLayerMasksOverride = prioritizedLayerMasks;
-                if (coneCastHit.collider != null)
-                {
-                    hitInfo = new MixedRealityRaycastHit(true, coneCastHit);
-                    return true;
-                }
-                else
-                {
-                    hitInfo = new MixedRealityRaycastHit();
-                    return false;
-                }
-            }
-
-            public bool OnSceneQuery(LayerMask[] prioritizedLayerMasks, bool focusIndividualCompoundCollider, out GameObject hitObject, out Vector3 hitPoint, out float hitDistance)
-            {
-                PrioritizedLayerMasksOverride = prioritizedLayerMasks;
-                if (coneCastHit.collider != null)
-                {
-                    hitObject = coneCastHit.collider.gameObject;
-                    hitPoint = coneCastHit.point;
-                    hitDistance = coneCastHit.distance;
-                    return true;
-                }
-                else
-                {
-                    hitObject = null;
-                    hitPoint = Vector3.zero;
-                    hitDistance = Mathf.Infinity;
-                    return false;
-                }
-            }
-
-            private static readonly ProfilerMarker OnPostSceneQueryPerfMarker = new ProfilerMarker("[MRTK] InternalConeCastPointer.OnPostSceneQuery");
-
-            /// <inheritdoc />
-            public override void OnPostSceneQuery()
-            {
-                using (OnPostSceneQueryPerfMarker.Auto())
-                {
-                    if (isDown)
-                    {
-                        CoreServices.InputSystem.RaisePointerDragged(this, MixedRealityInputAction.None, Controller.ControllerHandedness);
-                    }
-                }
-            }
-
-            /// <inheritdoc />
-            public override void OnPreCurrentPointerTargetChange() { }
-
-            /// <inheritdoc />
-            public override Vector3 Position => gazeTransform.position;
-
-            /// <inheritdoc />
-            public override Quaternion Rotation => gazeTransform.rotation;
-
-
-            /// <inheritdoc />
-            public override void Reset()
-            {
-                Controller = null;
-                BaseCursor = null;
-                IsActive = false;
-                IsFocusLocked = false;
-            }
-
-            /// <summary>
-            /// Press this pointer. This sends a pointer down event across the input system.
-            /// </summary>
-            /// <param name="mixedRealityInputAction">The input action that corresponds to the pressed button or axis.</param>
-            /// <param name="handedness">Optional handedness of the source that pressed the pointer.</param>
-            public void RaisePointerDown(MixedRealityInputAction mixedRealityInputAction, Handedness handedness = Handedness.None, IMixedRealityInputSource inputSource = null)
-            {
-                isDown = true;
-                currentHandedness = handedness;
-                currentInputSource = inputSource;
-                CoreServices.InputSystem?.RaisePointerDown(this, mixedRealityInputAction, handedness, inputSource);
-            }
-
-            /// <summary>
-            /// Release this pointer. This sends pointer clicked and pointer up events across the input system.
-            /// </summary>
-            /// <param name="mixedRealityInputAction">The input action that corresponds to the released button or axis.</param>
-            /// <param name="handedness">Optional handedness of the source that released the pointer.</param>
-            public void RaisePointerUp(MixedRealityInputAction mixedRealityInputAction, Handedness handedness = Handedness.None, IMixedRealityInputSource inputSource = null)
-            {
-                isDown = false;
-                currentHandedness = Handedness.None;
-                currentInputSource = null;
-                CoreServices.InputSystem?.RaisePointerClicked(this, mixedRealityInputAction, 0, handedness, inputSource);
-                CoreServices.InputSystem?.RaisePointerUp(this, mixedRealityInputAction, handedness, inputSource);
-            }
-        }
+        #endregion InternalGazePointer Classgaze
 
         #region MonoBehaviour Implementation
 
@@ -627,13 +422,23 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 //{
                 //    GazeCursor.SetVisibility(!internalPointer.IsFocusLocked);
                 //}
-                
+
+                // Handle toggling the input source's SourceType based on the current eyetracking mode 
+                if (IsEyeTrackingEnabledAndValid)
+                {
+                    gazeInputSource.SourceType = InputSourceType.Eyes;
+                }
+                else
+                {
+                    gazeInputSource.SourceType = InputSourceType.Head;
+                }
+
                 // If flagged to do so (setCursorInvisibleWhenFocusLocked) and active (IsInteractionEnabled), set the visibility to !IsFocusLocked,
                 // but don't touch the visibility when not active or not flagged.
-                if (setCursorInvisibleWhenFocusLocked && internalConeCastPointer != null &&
-                    internalConeCastPointer.IsInteractionEnabled && GazeCursor != null && internalConeCastPointer.IsFocusLocked == GazeCursor.IsVisible)
+                if (setCursorInvisibleWhenFocusLocked && coneCastPointer != null &&
+                coneCastPointer.IsInteractionEnabled && GazeCursor != null && coneCastPointer.IsFocusLocked == GazeCursor.IsVisible)
                 {
-                    GazeCursor.SetVisibility(!internalConeCastPointer.IsFocusLocked);
+                    GazeCursor.SetVisibility(!coneCastPointer.IsFocusLocked);
                 }
             }
         }
@@ -731,7 +536,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 if (eventData.InputSource.Pointers[i].PointerId == GazePointer.PointerId)
                 {
                     //internalPointer.RaisePointerUp(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
-                    internalConeCastPointer.RaisePointerUp(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
+                    coneCastPointer.RaisePointerUp(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
                     return;
                 }
             }
@@ -744,7 +549,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 if (eventData.InputSource.Pointers[i].PointerId == GazePointer.PointerId)
                 {
                     //internalPointer.RaisePointerDown(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
-                    internalConeCastPointer.RaisePointerDown(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
+                    coneCastPointer.RaisePointerDown(eventData.MixedRealityInputAction, eventData.Handedness, eventData.InputSource);
                     return;
                 }
             }
@@ -768,7 +573,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 Debug.Assert(gazeTransform != null, "No gaze transform to raycast from!");
 
                 //internalPointer = new InternalGazePointer(this, "Gaze Pointer", null, raycastLayerMasks, maxGazeCollisionDistance, gazeTransform, stabilizer);
-                internalConeCastPointer = new InternalConeCastPointer(this, "Gaze Pointer", null, raycastLayerMasks, maxGazeCollisionDistance, gazeTransform, stabilizer);
+                coneCastPointer = new GazeConePointer(this, "Gaze Pointer", null, raycastLayerMasks, maxGazeCollisionDistance, gazeTransform, stabilizer);
 
 
                 if ((GazeCursor == null) &&
@@ -782,7 +587,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
                 // Initialize gaze pointer
                 //internalPointer.IsTargetPositionLockedOnFocusLock = lockCursorWhenFocusLocked;
-                internalConeCastPointer.IsTargetPositionLockedOnFocusLock = lockCursorWhenFocusLocked;
+                coneCastPointer.IsTargetPositionLockedOnFocusLock = lockCursorWhenFocusLocked;
             }
         }
